caltech: Fix warning in bool.c. bool.c: In function bool_info: bool.c:1182:30: warning: format %d expects argument of type int, but argument 2 has type long unsigned int [-Wformat=]  1182 |   printf ("memory per node: %d\n", sizeof(bool_t));       |                             ~^     ~~~~~~~~~~~~~~       |                              |     |       |                              int   long unsigned int       |                             %ld
add cast to int.
Do not take the advise of the warning as it is system-specific.
C99 does offer %z for size_t, granted.

+++ caltech-other/bool/src/bool.c
@@ -1179,7 +1179,7 @@ void bool_print (bool_t *b)
  *-----------------------------------------------------------------------*/
 void bool_info (BOOL_T *B)
 {
-  printf ("memory per node: %d\n", sizeof(bool_t));
+  printf ("memory per node: %d\n", (int)sizeof(bool_t));
   printf ("max. num. of vars: %ld\n", (1UL<<(sizeof(bool_var_t)*8-1)));
   printf ("init. hashtable size: %d\n", HASH_BLOCK);
   printf ("var. block size: %d\n", VAR_BLOCK);
